got reid of players componenet, condensing the api data into the game componenet
