+++ client/src/components/BlackjackGame/style.css
@@ -1,0 +1,32 @@
+#house,
+#player1 {
+    text-align: center;
+    margin: 10px;
+    background-color: black;
+}
+
+#cardOneHouse,
+#cardTwoHouse,
+#cardOnePlayer1,
+#cardTwoPlayer1,
+.hitCardHouse,
+.hitCardPlayer1 {
+    width:100px;
+    height:150px;
+}
+
+#pointDisplay,
+#pointDisplayHousePlayer,
+#scoresDisplay,
+#scoresDisplayHousePlayer {
+    margin: 5px;
+    padding: 5px;
+
+}
+
+#roundOver,
+#winnerDisplay {
+    margin: 10px;
+    padding: 10px;
+
+}
