+++ client/src/components/BlackjackScoreCard/index.js
@@ -7,6 +7,9 @@ function BlackjackScoreCard() {
     return (
         <Card id="scoreCard" className="card mx-auto pt-3 pb-3">
             <p>This card will show at end of round.</p>
+            {/* winner declaration */}
+            {/* final points */}
+            {/* restart game button - props */}
         </Card>
     );
 }
