+++ client/src/components/BlackjackTable/index.js
@@ -4,23 +4,16 @@ import { Container } from 'reactstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import BlackjackDirections from "../BlackjackDirections";
 import BlackjackGame from "../BlackjackGame";
-// import BlackjackScoreCard from "../BlackjackScoreCard";
 
 function BlackjackTable() {
     const [displayDirections, setDisplayDirections] = useState(true);
     const [startGame, setStartGame] = useState(false);
-    // const [displayScoreCard, setDisplayScoreCard] = useState(false);
 
     function onStart() {
         console.log("clicking start btn");
         setStartGame(true);
         setDisplayDirections(false);
     }
-
-    // function endRound() {
-    //     console.log("end of round triggered");
-    //     setDisplayScoreCard(true);
-    // }
 
     return (
         <Container id="blackjackTable">
@@ -30,9 +23,6 @@ function BlackjackTable() {
             && <BlackjackDirections 
             onStart={onStart}
             />}
-            {/* {displayScoreCard
-            && <BlackjackScoreCard /> 
-            } */}
             {startGame 
             && <BlackjackGame /> 
             }
